reduce3 working
