+++ nd4j-backends/nd4j-api-parent/nd4j-api/src/main/java/org/nd4j/linalg/api/ops/BaseOp.java
@@ -66,8 +66,6 @@ public abstract class BaseOp implements Op {
                 }
                 return retBuff;
             }
-
-
         }
         return null;
     }
