+++ nd4j-backends/nd4j-backend-impls/nd4j-cuda-7.5/src/main/java/org/nd4j/linalg/jcublas/ops/executioner/JCudaExecutioner.java
@@ -29,6 +29,7 @@ import org.nd4j.linalg.api.complex.IComplexNumber;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.api.ops.*;
 import org.nd4j.linalg.api.ops.executioner.DefaultOpExecutioner;
+import org.nd4j.linalg.api.ops.impl.accum.Variance;
 import org.nd4j.linalg.api.ops.impl.transforms.arithmetic.CopyOp;
 import org.nd4j.linalg.factory.Nd4j;
 import org.nd4j.linalg.jcublas.buffer.AddressRetriever;
@@ -76,7 +77,7 @@ public class JCudaExecutioner extends DefaultOpExecutioner {
         }
 
 
-        if(dimension[0] == Integer.MAX_VALUE) {
+        if(dimension[0] == Integer.MAX_VALUE|| dimension[0] == 0) {
             if(op.x() instanceof IComplexNDArray) {
                 return Nd4j.scalar(execAndReturn(op).getFinalResultComplex());
             }
@@ -128,8 +129,10 @@ public class JCudaExecutioner extends DefaultOpExecutioner {
             }
 
             //nothing to reduce
-            if(ArrayUtil.prod(retShape) == op.x().length())
+            if(ArrayUtil.prod(retShape) == op.x().length()) {
                 return op.x();
+            }
+
             invoke(op,dimension);
             if(op.z() == null)
                 throw new IllegalStateException("No result set");
@@ -288,7 +291,6 @@ public class JCudaExecutioner extends DefaultOpExecutioner {
 
 
     private CudaContext invoke(BroadcastOp op) {
-        System.out.println("OPZF Broadcast");
         long x = AtomicAllocator.getInstance().getDevicePointer(op.x()).getNativePointer();
         long xShapeInfo = AddressRetriever.retrieveDeviceAddress(op.x().shapeInfoDataBuffer());
         long[] xShapeInfoHostPointer = new long[]{AddressRetriever.retrieveHostAddress(op.x().shapeInfoDataBuffer()), AtomicAllocator.getInstance().getCudaContext().getOldStream().getNativePointer()};
@@ -324,7 +326,7 @@ public class JCudaExecutioner extends DefaultOpExecutioner {
         long xShapeInfo = AddressRetriever.retrieveDeviceAddress(op.x().shapeInfoDataBuffer());
         long extraArgs = op.extraArgs() != null ? AddressRetriever.retrieveDeviceAddress(op.extraArgsDataBuff()) : 0;
         long[] xShapeInfoHostPointer = new long[]{AddressRetriever.retrieveHostAddress(op.x().shapeInfoDataBuffer()), AtomicAllocator.getInstance().getCudaContext().getOldStream().getNativePointer()};
-        if(op.z().isScalar()) {
+        if(op.z().isScalar() || dimension == null || dimension[0] == Integer.MAX_VALUE) {
             if(op.x().data().dataType() == DataBuffer.Type.DOUBLE) {
                 double result = nativeOps.execIndexReduceScalarDouble(xShapeInfoHostPointer,op.opNum(),x,xShapeInfo,extraArgs);
                 op.setFinalResult((int) result);
@@ -363,77 +365,63 @@ public class JCudaExecutioner extends DefaultOpExecutioner {
 
 
     private CudaContext invoke(Accumulation op, int[] dimension) {
+        // dimension is ALWAYS null here.
         if (dimension == null) dimension = new int[] {Integer.MAX_VALUE};
         CudaContext ctx = null;
         long[] xShapeInfoHostPointer = new long[]{AddressRetriever.retrieveHostAddress(op.x().shapeInfoDataBuffer()), AtomicAllocator.getInstance().getCudaContext().getOldStream().getNativePointer()};
-        if(op.y() != null) {
-            long x = AtomicAllocator.getInstance().getDevicePointer(op.x()).getNativePointer();
-            long xShapeInfo = AddressRetriever.retrieveDeviceAddress(op.x().shapeInfoDataBuffer());
-            long y = AtomicAllocator.getInstance().getDevicePointer(op.y()).getNativePointer();
-            long yShapeInfo = AddressRetriever.retrieveDeviceAddress(op.y().shapeInfoDataBuffer());
-            long extraArgs = op.extraArgs() != null ? AddressRetriever.retrieveDeviceAddress(op.extraArgsDataBuff()) : 0;
-            if(op.z().isScalar()) {
-                if(op.x().data().dataType() == DataBuffer.Type.DOUBLE) {
-                    double result = nativeOps.execReduce3ScalarDouble(
-                            xShapeInfoHostPointer,
-                            op.opNum()
-                            ,x,
-                            xShapeInfo,
-                            extraArgs,
-                            y,
-                            yShapeInfo);
-                    op.setFinalResult(result);
-                }
-                else {
-                    float result = nativeOps.execReduce3ScalarFloat(xShapeInfoHostPointer,
-                            op.opNum()
-                            , x,
-                            xShapeInfo,
-                            extraArgs,
-                            y,
-                            yShapeInfo);
-                    op.setFinalResult(result);
-                }
+        long x = AtomicAllocator.getInstance().getDevicePointer(op.x()).getNativePointer();
+        long xShapeInfo = AddressRetriever.retrieveDeviceAddress(op.x().shapeInfoDataBuffer());
+        long extraArgs = op.extraArgs() != null ? AddressRetriever.retrieveDeviceAddress(op.extraArgsDataBuff()) : 0;
+
+        if (op.x().data().dataType() == DataBuffer.Type.DOUBLE) {
+            if(op instanceof Variance) {
+                double result = nativeOps.execSummaryStatsScalarDouble(
+                        xShapeInfoHostPointer,
+                        op.opNum(),
+                        x
+                        ,xShapeInfo,extraArgs, true);
+                op.setFinalResult(result);
+            } else if (op.y() != null) {
+                long y = AtomicAllocator.getInstance().getDevicePointer(op.y()).getNativePointer();
+                long yShapeInfo = AddressRetriever.retrieveDeviceAddress(op.y().shapeInfoDataBuffer());
+                double result = nativeOps.execReduce3ScalarDouble(xShapeInfoHostPointer,
+                        op.opNum()
+                        , x,
+                        xShapeInfo,
+                        extraArgs,
+                        y,
+                        yShapeInfo);
+                op.setFinalResult(result);
+            } else {
+                double result = nativeOps.execReduceScalarDouble(xShapeInfoHostPointer, op.opNum(), x, xShapeInfo, extraArgs);
+                op.setFinalResult(result);
             }
-            else {
-                long z = AtomicAllocator.getInstance().getDevicePointer(op.z()).getNativePointer();
-                long zShapeInfo = AddressRetriever.retrieveDeviceAddress(op.z().shapeInfoDataBuffer());
-                if(op.x().data().dataType() == DataBuffer.Type.DOUBLE) {
-                    nativeOps.execReduce3Double(xShapeInfoHostPointer,op.opNum(),x,xShapeInfo,extraArgs,y,yShapeInfo,z,zShapeInfo);
-                }
-                else {
-                    nativeOps.execReduce3Float(xShapeInfoHostPointer,op.opNum(),x,xShapeInfo,extraArgs,y,yShapeInfo,z,zShapeInfo);
-                }
+        } else {
+            if(op instanceof Variance) {
+                float result = nativeOps.execSummaryStatsScalarFloat(
+                        xShapeInfoHostPointer,
+                        op.opNum(),
+                        x
+                        ,xShapeInfo,extraArgs, true);
+                op.setFinalResult(result);
+            } else if (op.y() != null) {
+                long y = AtomicAllocator.getInstance().getDevicePointer(op.y()).getNativePointer();
+                long yShapeInfo = AddressRetriever.retrieveDeviceAddress(op.y().shapeInfoDataBuffer());
+                float result = nativeOps.execReduce3ScalarFloat(xShapeInfoHostPointer,
+                        op.opNum()
+                        , x,
+                        xShapeInfo,
+                        extraArgs,
+                        y,
+                        yShapeInfo);
+                op.setFinalResult(result);
+            } else {
+                float result = nativeOps.execReduceScalarFloat(xShapeInfoHostPointer, op.opNum(), x, xShapeInfo, extraArgs);
+                op.setFinalResult(result);
             }
         }
-        else {
-            long x = AtomicAllocator.getInstance().getDevicePointer(op.x()).getNativePointer();
-            long xShapeInfo = AddressRetriever.retrieveDeviceAddress(op.x().shapeInfoDataBuffer());
-            long extraArgs = op.extraArgs() != null ? AddressRetriever.retrieveDeviceAddress(op.extraArgsDataBuff()) : 0;
-            if(op.z().isScalar()) {
-                if(op.x().data().dataType() == DataBuffer.Type.DOUBLE) {
-                    double result = nativeOps.execReduceScalarDouble(xShapeInfoHostPointer, op.opNum(), x, xShapeInfo,extraArgs);
-                    op.setFinalResult(result);
-                }
-                else {
-                    float result = nativeOps.execReduceScalarFloat(xShapeInfoHostPointer, op.opNum(), x, xShapeInfo, extraArgs);
-                    op.setFinalResult(result);
-                }
-            }
-            else {
-                long z = AtomicAllocator.getInstance().getDevicePointer(op.z()).getNativePointer();
-                long zShapeInfo = AddressRetriever.retrieveDeviceAddress(op.z().shapeInfoDataBuffer());
-                long dimensionPointer = AddressRetriever.retrieveDeviceAddress(Nd4j.createBuffer(dimension));
-                if(op.x().data().dataType() == DataBuffer.Type.DOUBLE) {
-                    nativeOps.execReduceDouble(xShapeInfoHostPointer,op.opNum(),x,xShapeInfo,extraArgs,z,zShapeInfo,dimensionPointer,dimension.length);
-                }
-                else {
-                    nativeOps.execReduceFloat(xShapeInfoHostPointer, op.opNum(), x, xShapeInfo, extraArgs, z, zShapeInfo,dimensionPointer,dimension.length);
 
-                }
-            }
 
-        }
         if (op.x() != null)
             allocator.tackDevice(op.x());
         if (op.y() != null)
@@ -444,10 +432,6 @@ public class JCudaExecutioner extends DefaultOpExecutioner {
         // we notify allocator that op.Z was modified on device side
         if (op.z() != null)
             allocator.tickDeviceWrite(op.z());
-
-        if (op.z().isScalar())
-            op.setFinalResult(op.z().getDouble(0));
-
 
         return ctx;
     }
@@ -479,6 +463,7 @@ public class JCudaExecutioner extends DefaultOpExecutioner {
     }
 
     private CudaContext invoke(TransformOp op) {
+//        System.out.println("OpName: ["+op.getClass().getSimpleName()+"]; OpCode: ["+op.opNum()+"]");
         long x = AtomicAllocator.getInstance().getDevicePointer(op.x()).getNativePointer();
         long xShapeInfo = AddressRetriever.retrieveDeviceAddress(op.x().shapeInfoDataBuffer());
         long extraArgs = op.extraArgs() != null ? AddressRetriever.retrieveDeviceAddress(op.extraArgsDataBuff()) : 0;
@@ -503,7 +488,6 @@ public class JCudaExecutioner extends DefaultOpExecutioner {
             }
             else {
                 nativeOps.execTransformFloat(xShapeInfoHostPointer,op.opNum(),x,xShapeInfo,z,zShapeInfo,extraArgs);
-
             }
         }
         if (op.x() != null) allocator.tackDevice(op.x());
